refactor(work): split FlowView.tsx into Tabs/Toolbar/Chart/InspectPanel (CHAOS-1572)
Splits src/components/work/FlowView.tsx (624 LOC) into a thin composition
root plus focused sub-components co-located under src/components/work/FlowView/:

  - Tabs.tsx (88 LOC)        — tab nav, keyboard navigation, tabRefs
  - Toolbar.tsx (37 LOC)     — chart-section header + ChartTypeToggle
  - Chart.tsx (152 LOC)      — chart variants (treemap, sunburst, area, sankey)
  - InspectPanel.tsx (133 LOC) — selection details, evidence/flame links
  - useFlowHandlers.ts (137 LOC) — click-handler hook (keeps root <=250 LOC)

FlowView.tsx (the composition root) is now 233 LOC; every sub-file is well
under the 250-LOC acceptance threshold.

Filters.tsx omitted — FlowView has no filter JSX; the filters prop is
passed through from the parent unchanged. Final split: Tabs, Toolbar,
Chart, InspectPanel (4 sub-components + 1 helper hook, matches the
ticket's count of 4 sub-files plus a small extracted hook for handler
state to satisfy the per-file LOC budget).

Verification:
  npx tsc --noEmit                       PASS (clean)
  npm run lint                           PASS (only pre-existing
                                              warning in sentry/scrubber.ts)
  npx vitest run src/components/work     PASS (17 tests, 3 files)

No behavior change. No prop API change for parent callers.
